+++ butakovskij.egor/T2/StreamGuard.h
@@ -2,11 +2,17 @@
 #define STREAMGUARD_H
 
 #include <ios>
+#include <ostream>
 
 class StreamGuard {
 public:
-    explicit StreamGuard(std::ostream& os);
-    ~StreamGuard();
+    explicit StreamGuard(std::ostream& os)
+    : stream(os), flags(os.flags()), fill(os.fill()), precision(os.precision()) {}
+    ~StreamGuard(){
+        stream.flags(flags);
+        stream.fill(fill);
+        stream.precision(precision);
+    }
 
     StreamGuard(const StreamGuard&) = delete;
     StreamGuard& operator=(const StreamGuard&) = delete;
@@ -16,6 +22,8 @@ private:
     std::ios::fmtflags flags;
     char fill;
     std::streamsize precision;
+
 };
+
 
 #endif
